Split Dynamic Worker reuse executor tests under the file-size ratchet
The new invoke coverage pushed executor.node.test.ts over 2000 lines.
