+++ packages/worker/src/mcp/executor.node.test.ts
@@ -1179,96 +1179,6 @@ test('createExecuteExecutor records one unique Dynamic Worker day per worker id'
 	expect(dataPoints[2]?.blobs?.[7]).toBe('hit')
 })
 
-test('createExecuteExecutor records privacy-safe Dynamic Worker reuse on every LOADER invoke', async () => {
-	const dataPoints: Array<AnalyticsEngineDataPoint> = []
-	const meter = createInMemoryUserMeterEnv()
-	const usageBindings = {
-		...meter.env,
-		USAGE_EVENTS: {
-			writeDataPoint(point?: AnalyticsEngineDataPoint) {
-				if (point) dataPoints.push(point)
-			},
-		},
-	}
-	const exports = createExecutorTestExports()
-	const providers = [{ name: 'kody', fns: {} }]
-	const sourceMarker = 'UNIQUE_SOURCE_MARKER_reuse_metrics'
-	const paramMarker = 'UNIQUE_PARAM_MARKER_reuse_metrics'
-	const source = `async () => "${sourceMarker}"`
-
-	const firstLoader = createFakeWorkerLoader()
-	await createExecuteExecutor({
-		env: {
-			...createExecutorTestEnv(firstLoader.loader),
-			...usageBindings,
-		} as Env,
-		exports,
-		gatewayProps: createGatewayProps('usage-user-reuse'),
-		recordExecuteUsage: false,
-		surface: 'job',
-		executeShape: 'glue',
-	}).execute(source, providers, {
-		params: { token: paramMarker },
-	})
-
-	const miss = dataPoints.find(
-		(point) => point.blobs?.[1] === 'dynamic_worker_invoke',
-	)
-	expect(miss?.blobs?.[5]).toBe('job')
-	expect(miss?.blobs?.[6]).toBe('glue')
-	expect(miss?.blobs?.[7]).toBe('miss')
-	expect(miss?.blobs?.[8]).toBe('true')
-	expect(miss?.doubles?.[0]).toBeGreaterThanOrEqual(0)
-	expect(miss?.doubles?.[3]).toBeGreaterThan(0)
-	const missCodeChars = miss?.doubles?.[3] ?? 0
-
-	const secondLoader = createFakeWorkerLoader()
-	await createExecuteExecutor({
-		env: {
-			...createExecutorTestEnv(secondLoader.loader),
-			...usageBindings,
-		} as Env,
-		exports,
-		gatewayProps: createGatewayProps('usage-user-reuse'),
-		recordExecuteUsage: false,
-		surface: 'job',
-		executeShape: 'glue',
-	}).execute(source, providers, {
-		params: { token: `${paramMarker}-2` },
-	})
-
-	const invokes = dataPoints.filter(
-		(point) => point.blobs?.[1] === 'dynamic_worker_invoke',
-	)
-	expect(invokes).toHaveLength(2)
-	expect(invokes[1]?.blobs?.[7]).toBe('hit')
-	expect(invokes[1]?.blobs?.[8]).toBe('true')
-	expect(invokes[1]?.doubles?.[3]).toBe(missCodeChars)
-
-	const thirdLoader = createFakeWorkerLoader()
-	await createExecuteExecutor({
-		env: {
-			...createExecutorTestEnv(thirdLoader.loader),
-			...usageBindings,
-		} as Env,
-		exports,
-		gatewayProps: createGatewayProps('usage-user-reuse'),
-		recordExecuteUsage: false,
-		surface: 'job',
-	}).execute(source, providers)
-
-	expect(
-		dataPoints.filter((point) => point.blobs?.[1] === 'dynamic_worker_invoke'),
-	).toHaveLength(3)
-	expect(dataPoints.at(-1)?.blobs?.[7]).toBe('hit')
-	expect(dataPoints.at(-1)?.blobs?.[8]).toBe('false')
-
-	const serialized = JSON.stringify(dataPoints)
-	expect(serialized).not.toContain(sourceMarker)
-	expect(serialized).not.toContain(paramMarker)
-	expect(serialized).not.toContain(source)
-})
-
 test('createExecuteExecutor defers unique-worker-day and first-execute stamp via waitUntil', async () => {
 	const dataPoints: Array<AnalyticsEngineDataPoint> = []
 	const activationStampWrites: Array<string> = []
